% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbfm_functions.R
\name{corr.est}
\alias{corr.est}
\title{Estimate the gene-gene correlation matrix from MCMC samples}
\usage{
corr.est(hbfm.list)
}
\arguments{
\item{hbfm.list}{list where each element contains an hbfm.fit-class object; each element of the list contains an object from a different MCMC chain}
}
\value{
hbfm.corr-class object containing:

 \itemize{ 
  \item{corr}{: symmetric estimated correlation matrix generated from MCMC samples}
  \item{CI.low}{: symmetric matrix representing the lower bound of the 95\% credible intervals (CI) generated from the MCMC samples}
  \item{CI.upp}{: symmetric matrix representing the upper bound of the 95\% credible intervals (CI) generated from the MCMC samples}
  \item{CI.eval}{: symmetric logical matrix generated from 95\% credible intervals (CI) from MCMC samples; TRUE means 95\% CI does not include 0 (i.e., significant correlation)}
  \item{p.val}{: symmetric matrix consisting of approximate "p-values" generated from MCMC samples}
}
}
\description{
Function used to combine and analyze MCMC results from one or more sets of
correlation samples generated by the \code{hbfm.fit} function.
}
\details{
The correlation estimate is determined by calculating the average of the posterior samples for each gene-gene pair.

To determine whether the correlation is significant, a 95\% credible interval (CI) is determined from the posterior samples. If the CI
includes 0, the correlation is deemed to be non-significant and the "CI.eval" element is FALSE. If the CI does not include 0, the correlation is
considered significant and the "CI.eval" element is TRUE.

An alternative measurement of significance is the approximate "p-value" calculation, which is output in the "p.val" matrix. This "p-value" is determined 
by finding the smallest "a" value such that the 100(1-a)\% CI contains 0. The corresponding "a" value represents the proportion of posterior distribution 
that is outside the smallest credible interval that contains 0.

In most cases a significant 95\% CI from "CI.eval" will correspond with an approximate "p-value" < 0.05 from "p.val".
}
\examples{
\dontrun{
## Load dataset
data(gene.mat)

## Run stochastic EM first
## Consider F=5 factors
fit1 <- stoc.em(Y=gene.mat, Fac = 5)

## Run MCMC sampler with initial parameter values from stoc.em
fit.res1 <- hbfm.fit(fit1)

## Obtain estimated gene-gene correlations from MCMC samples
fit.corr <- corr.est(list(fit.res1))
print(fit.corr)

}

}
