% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbfm_functions.R
\docType{data}
\name{gene.mat}
\alias{gene.mat}
\title{Example gene expression matrix}
\format{A matrix with 10 rows and 100 columns}
\usage{
gene.mat
}
\description{
An example dataset consisting of gene expression counts for G = 10 genes (rows)
across N = 100 cells (columns)
}
\examples{
data(gene.mat)
rownames(gene.mat)
colnames(gene.mat)

}
\keyword{datasets}
