% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbfm_functions.R
\name{hbfm.DIC}
\alias{hbfm.DIC}
\title{Calculate DIC of hbfm}
\usage{
hbfm.DIC(hbfm.list)
}
\arguments{
\item{hbfm.list}{list where each element contains an hbfm.fit-class object; each element of the list contains an object from a different MCMC chain}
}
\value{
DIC of hbfm with \code{Fac} number of factors
}
\description{
Function to calculate the Deviance Information Criterion (DIC) from \code{hbfm.fit} output.
}
\details{
For each of the final \code{M.save} iterations of the MCMC, \code{hbfm.fit} calculates the log marginal likelihood, log(ML).
 
The DIC is calculated with the formula: \eqn{DIC = mean(D) + pD}; where \eqn{D = -2*log(ML)} and \eqn{pD = var(D)/2}.
}
\examples{
\dontrun{
## Load dataset
data(gene.mat)

## Run stochastic EM first
## Consider F=5 factors
fit1 <- stoc.em(Y=gene.mat, Fac = 5)

## Run MCMC sampler with initial parameter values from stoc.em
fit.res1 <- hbfm.fit(fit1)

## Obtain estimated gene-gene correlations from MCMC samples
hbfm.DIC(list(fit.res1))

}
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., & Rubin, D. B. (2004). \emph{Bayesian Data Analysis: 2nd Edition}. Chapman and Hall/CRC.
}
