% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbfm_functions.R
\name{hbfm.fit}
\alias{hbfm.fit}
\title{Hierarchical Bayesian factor model MCMC sampler}
\usage{
hbfm.fit(stoc.em.param, M = 4000, M.save = 1000, M.ll.seq = 10,
  H = 50, phi.scale = 0.25, par.samp = FALSE, seed = 123,
  verbose = FALSE)
}
\arguments{
\item{stoc.em.param}{object of hbfm.par-class created by the \code{stoc.em} function}

\item{M}{total number of MCMC iterations}

\item{M.save}{number of iterations to be used for correlation estimation; the final \code{M.save} iterations from
the \code{M} number of MCMC iterations will be saved for further analysis}

\item{M.ll.seq}{intervals for calculating marginal log-likelihood before final \code{M.save} iterations; used 
to reduce computational time}

\item{H}{number of lambda draws for marginal log-likelihood calculation}

\item{phi.scale}{lognormal scale parameter used when drawing candidate values of phi in
Metropolis-Hastings step; can be used to adjust acceptance rate of phi samples}

\item{par.samp}{if TRUE, samples of each parameter in \code{hbfm.fit} for each MCMC iteration are saved; if FALSE
only the calculated parameters of correlation and marginal log-likelihood are saved}

\item{seed}{seed for random number generation}

\item{verbose}{if TRUE, \code{hbfm.fit} status is displayed at every 100th iteration}
}
\value{
hbfm.fit-class object containing:

 \itemize{ 
  \item{Y}{: data.frame or matrix of gene expression counts}
  \item{Fac}{: number of factors considered in the model}
  \item{samples}{: samples of estimated parameters and/or calculated parameters from MCMC iterations}
  \itemize{
  \item{mll}{: calculated marginal log-likelihoods}
  \item{corr}{: samples of gene-gene correlation matrix}
  \item{beta}{: samples of beta parameters; included only when \code{par.samp = TRUE}}
  \item{theta}{: samples of theta parameters; included only when \code{par.samp = TRUE}}
  \item{alpha}{: samples of alpha parameters; included only when \code{par.samp = TRUE}}
  \item{lambda}{: samples of lambda parameters; included only when \code{par.samp = TRUE}}
  \item{phi}{: samples of phi parameters; included only when \code{par.samp = TRUE}}
  \item{h1}{: samples of location hyperparameter of lognormal prior for phi; included only when \code{par.samp = TRUE}}
  \item{h2}{: samples of scale hyperparameter of lognormal prior for phi; included only when \code{par.samp = TRUE}}
  }
}
}
\description{
Function used to implement the MCMC sampler for the hierarchical Bayesian factor model (HBFM) 
defined in the "A sparse Bayesian factor model for the construction of gene co-expression 
networks from single-cell RNA sequencing count data" manuscript.
}
\examples{
\dontrun{
## Load dataset
data(gene.mat)

## Run stochastic EM first
## Consider F=5 factors
fit1 <- stoc.em(Y=gene.mat, Fac = 5)

## Run MCMC sampler with initial parameter values from stoc.em
fit.res1 <- hbfm.fit(fit1)
}


}
