% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbfm_functions.R
\name{rank.chains}
\alias{rank.chains}
\title{Rank hbfm chains}
\usage{
rank.chains(hbfm.list)
}
\arguments{
\item{hbfm.list}{list where each element contains an hbfm.fit-class object; each element of the list contains an object from a different MCMC chain}
}
\value{
data.frame that orders the input hbfm.fit-class objects based on average marginal log-likelihood; the numbers in the chain.num column
represent the elements (chains) from hbfm.list
}
\description{
Function to rank the MCMC chains by average marginal log-likelihood.
}
\details{
The \code{rank.chains} function is used to rank the input chains based on average marginal log-likelihood in order to identify low performing chains.
}
\examples{
\dontrun{
## Load dataset
data(gene.mat)

## Run stochastic EM first
## Consider F=5 factors
fit1 <- stoc.em(Y=gene.mat, Fac = 5)

## Run MCMC sampler with initial parameter values from stoc.em
fit.res1 <- hbfm.fit(fit1)

## Run a second chain
fit2 <- stoc.em(Y=gene.mat, Fac = 5, seed = 234)
fit.res2 <- hbfm.fit(fit2, seed = 234)

## Obtain estimated gene-gene correlations from MCMC samples
rank.chains(list(fit.res1, fit.res2))

}

}
