% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbfm_functions.R
\name{stoc.em}
\alias{stoc.em}
\title{Stochastic EM algorithm for initial parameter estimation}
\usage{
stoc.em(Y, Fac, M.stoc = 2000, M.int = 100, M.eval = 200,
  M.ll.seq = 10, H = 50, seed = 123, verbose = FALSE)
}
\arguments{
\item{Y}{data.frame or matrix of gene expression counts where the rows correspond to genes and
columns correspond to cells; Y must contain integers and have row names}

\item{Fac}{number of factors to consider in the model; only a single number is accepted}

\item{M.stoc}{total number of stochastic EM iterations}

\item{M.int}{initial number of MCMC draws before maximization}

\item{M.eval}{number of iterations to be used for parameter estimation; the final \code{M.eval} iterations from
the \code{M.stoc} number of stochastic EM iterations will be considered}

\item{M.ll.seq}{intervals for calculating marginal log-likelihood before final \code{M.eval} iterations; used 
to reduce computational time}

\item{H}{number of lambda draws for marginal log-likelihood calculation}

\item{seed}{seed for random number generation}

\item{verbose}{if TRUE, \code{stoc.em} status is displayed at every 100th iteration}
}
\value{
hbfm.par-class object containing:

\itemize{ 
  \item{Y}{: data.frame or matrix of gene expression counts}
  \item{Fac}{: number of factors considered in the model}
  \item{beta}{: initial beta parameter vector for input into \code{hbfm.fit}}
  \item{theta}{: initial theta parameter vector for input into \code{hbfm.fit}}
  \item{alpha}{: initial alpha parameter matrix for input into \code{hbfm.fit}}
  \item{lambda}{: initial lambda parameter matrix for input into \code{hbfm.fit}}
  \item{phi}{: initial phi parameter vector for input into \code{hbfm.fit}}
  \item{h1}{: location hyperparameter of lognormal prior for phi to input into \code{hbfm.fit}}
  \item{h2}{: scale hyperparameter of lognormal prior for phi to input into \code{hbfm.fit}}
  \item{mll}{: calculated marginal log-likelihoods at given iterations; used for convergence diagnostics}
}
}
\description{
Function used to implement the stochastic EM algorithm defined in the
"A sparse Bayesian factor model for the construction of gene co-expression 
networks from single-cell RNA sequencing count data" manuscript.
}
\details{
This algorithm should be used before \code{hbfm.fit} as it generates initial parameter values
for use in \code{hbfm.fit}.
}
\examples{
\dontrun{
## Load dataset
data(gene.mat)

## Run stochastic EM first
## Consider F=5 factors
fit1 <- stoc.em(Y=gene.mat, Fac = 5)

}

}
